\name{AMA2006}
\docType{data}
\alias{AMA2006}

\title{AMA2006 cell count table}

\description{
This dataset was obtained from a cambial activity and wood formation monitoring study. Data were collected in 2006 for 5 silver firs and 5 Scots pines grown in the same stand, located in Amance forest near Nancy (northeast France). \cr
Column "Tree" indicates the tree reference number, "Sp" indicates the tree species code, "DY" indicates the date of the sampling in day of year, RF indicates the radial file reference number. "nC" indicates the number of cells in the cambial zone, "nE" indicates the number of cells in the enlarging zone, "nL" indicates the number of cells in the maturing zone and "nM" indicates the number of mature cells. 
}

\usage{AMA2006}

\format{A data frame containing 8 columns and 940 rows}

\source{
Cyrille Rathgeber - LERFoB, UMR1092 - INRA Nancy (France) \cr
Email: \email{cyrille.rathgeber@nancy.inra.fr}
}

\references{
Rathgeber C.B.K., Longuetaud F. , Mothe F., Cuny H., Le Moguedec G. 2010. Phenology of wood formation: data processing, analysis and visualisation using R. Accepted in \emph{Dendrochronologia}.
}

\seealso{
\code{\link{averageRadialFiles}, \link{computeBootstrapTest}, \link{computeCriticalDates}, \link{computeRingCellNb}, \link{readXLSCellCountTable}, \link{fitGompertz}, \link{plotWoodFormationCalendar}, \link{plotWoodFormationDynamics}}
}

\note{
Last modifications: 19 July 2010
}

\keyword{datasets}
